deploy/requirements.txt
mlflow == 2.8.1
numpy == 1.24.3
pandas == 2.0.3
datasets == 2.15.0
scikit-learn == 1.3.2
fastapi == 0.104.1
uvicorn == 0.24.0.post1
torch == 1.12.1
transformers == 4.22.1
scikit-optimize == 0.9.0
xgboost == 2.0.2
python-dotenv == 0.21.0
nltk